Syscall fixes for 64-bit
- Fixed returns from syscalls in 64-bit mode; it's apparently required
  to explicitly do "sysretq" to return to 64-bit code.

- Implement reading of the tarball style init disk in root server; this
  change simplifies its implementation somewhat.

- Miscellaneous fixes in syscalls and printing of handles

- Made parsing of init scripts in root server more robust (with regard
  to reading the init script keywords)

- Fixed strlen (!) implementation breakage on 64-bit

+++ user/lib/libsystem/src/sys/irq.c
@@ -13,7 +13,7 @@ int IrqHandlerInstall(const uintptr_t irqNum, const uintptr_t threadHandle,
     }
 
     // do syscall
-    int err = __do_syscall3(SYS_ARCH_INSTALL_IRQ, irqNum, threadHandle, bits);
+    intptr_t err = __do_syscall3(SYS_ARCH_INSTALL_IRQ, irqNum, threadHandle, bits);
 
     if(err > 0) {
         *outHandle = err;
